Fixup CR handling, especially in RX window.
When you addch('\n') in curses, it does an implcit clreol().  Since
CR LF is often send at the end of lines, and CR moves to the beginning
of the current line, this results in a line being erased after it is
received.  We don't want to do that, so we handle \r and \n explicity.