Fix hfs buffer handling and DSP buffer reset
When we return from changing settings, or reverse mode, we need to
throw out the existing hfs buffer.  Pass a pointer into do_tx() and
handle it in there since the return value is already used to indicate
program termination.

For the DSP buffer, when we allocate a new buffer, reset head and tail
pointers to zero.

In the FIR filter, remove obsolete ret feeding, and only call memcpy()
if needed (rather than memcpy() zero lengths)