Experimental noise correction...
Each time a bit is sampled, note the current envelope of the "off"
frequency, and subtract it from successive samples of that frequency.

This may help with selective fading... ideally though, this would
be retroactive through a character instead of "real time" during them.

The noise levels are reset any time sync is lost.  The first character
must be received on both frequences for decode to start.