Make it multithreaded.
This commit adds an xmlrpc thread and an rtty_demod thread.  Likely
the next commit will add an AFSK thread.

The main thread becomes the input thread and blocks on wgetch() (and
sending AFSK data), the xml-rpc thread blocks on select(), and the
rtty_demod thread blocks on the sound card read().